refactor: Restructure models feature
