+++ application/ui/src/features/models/job-table/job-row-content.tsx
@@ -1,9 +1,9 @@
 import { Heading, Item, TabList, TabPanels, Tabs, View } from '@geti-ui/ui';
 
-import { SchemaTrainJob } from '../../api/openapi-spec';
-import { JobMetricsContent } from './metrics';
+import { SchemaTrainJob } from '../../../api/openapi-spec';
+import { JobMetricsContent } from '../metrics/metrics';
 
-import classes from './model-row-content.module.css';
+import classes from '../model-table/model-row-content.module.css';
 
 interface JobRowContentProps {
     job: SchemaTrainJob;
