+++ application/ui/src/features/models/model-table/model-row-content.module.css
