+++ application/ui/src/features/models/shared/box.module.css
