+++ application/ui/src/features/models/shared/box.tsx
