+++ application/ui/src/features/models/shared/collapsable-row.module.css
