+++ application/ui/src/features/models/shared/collapsable-row.tsx
