+++ application/ui/src/features/models/shared/duration.ts
@@ -1,0 +1,5 @@
+import { formatDuration } from '../../../utils';
+
+export const durationBetween = (start: string, end: string): string => {
+    return formatDuration(new Date(end).getTime() - new Date(start).getTime());
+};
