+++ application/ui/src/features/models/shared/split-badge.module.css
