+++ application/ui/src/features/models/shared/table-columns.ts
