+++ application/ui/src/routes/models/index.tsx
@@ -1,268 +1,1 @@
-import { useState } from 'react';
-
-import {
-    Button,
-    Content,
-    DialogContainer,
-    DialogTrigger,
-    Flex,
-    Heading,
-    IllustratedMessage,
-    Text,
-    View,
-} from '@geti-ui/ui';
-import { useQueryClient } from '@tanstack/react-query';
-import useWebSocket from 'react-use-websocket';
-
-import { $api, fetchClient } from '../../api/client';
-import { SchemaTrainJob as SchemaJob, SchemaModel } from '../../api/openapi-spec';
-import { notify } from '../../components/notification/notification.component';
-import { LogsDialog } from '../../features/logs/logs-dialog';
-import { useProjectId } from '../../features/projects/use-project';
-import { ReactComponent as EmptyIllustration } from './../../assets/illustration.svg';
-import { TrainingHeader, TrainingRow } from './job-table.component';
-import { ModelHeader, ModelRow } from './model-table.component';
-import { SchemaTrainJob, TrainModelDialog } from './train-model-dialog';
-
-const ModelList = ({
-    models,
-    jobs,
-    onRetrain,
-    onViewLogs,
-}: {
-    models: SchemaModel[];
-    jobs: SchemaJob[];
-    onRetrain: (model: SchemaModel) => void;
-    onViewLogs: (model: SchemaModel) => void;
-}) => {
-    const sortedModels = models.toSorted(
-        (a, b) => new Date(b.created_at!).getTime() - new Date(a.created_at!).getTime()
-    );
-
-    const jobsById = new Map(jobs.map((j) => [j.id, j]));
-
-    const { project_id } = useProjectId();
-    const deleteModelMutation = $api.useMutation('delete', '/api/models/{model_id}', {
-        meta: {
-            invalidates: [['get', '/api/projects/{project_id}/models', { params: { path: { project_id } } }]],
-        },
-    });
-
-    const deleteModel = (model: SchemaModel) => {
-        deleteModelMutation.mutate({ params: { path: { model_id: model.id! } } });
-    };
-
-    return (
-        <View
-            borderWidth='thin'
-            borderColor={'gray-200'}
-            borderBottomWidth='thin'
-            borderBottomColor={'gray-200'}
-            borderStartWidth='thin'
-            borderStartColor={'gray-200'}
-            borderEndWidth='thin'
-            borderEndColor={'gray-200'}
-        >
-            <ModelHeader />
-            {sortedModels.map((model) => (
-                <ModelRow
-                    key={model.id}
-                    model={model}
-                    trainingJob={model.train_job_id ? jobsById.get(model.train_job_id) : undefined}
-                    onDelete={() => deleteModel(model)}
-                    onRetrain={() => onRetrain(model)}
-                    onViewLogs={() => onViewLogs(model)}
-                />
-            ))}
-        </View>
-    );
-};
-
-const JobList = ({ jobs, onViewLogs }: { jobs: SchemaTrainJob[]; onViewLogs: (job: SchemaTrainJob) => void }) => {
-    const sortedJobs = jobs
-        .filter((m) => m.status !== 'completed')
-        .toSorted((a, b) => new Date(b.created_at!).getTime() - new Date(a.created_at!).getTime());
-
-    const interruptMutation = $api.useMutation('post', '/api/jobs/{job_id}:interrupt', {
-        meta: {
-            invalidates: [['get', '/api/jobs']],
-        },
-    });
-    const onInterrupt = (job: SchemaTrainJob) => {
-        if (job.id !== undefined) {
-            interruptMutation.mutate({
-                params: {
-                    path: {
-                        job_id: job.id,
-                    },
-                },
-            });
-        }
-    };
-
-    if (sortedJobs.length === 0) {
-        return <></>;
-    }
-
-    return (
-        <View marginBottom={'size-600'}>
-            <Heading level={4} marginBottom={'size-100'}>
-                Current Training
-            </Heading>
-
-            <TrainingHeader />
-            {sortedJobs.map((job) => (
-                <TrainingRow
-                    key={job.id}
-                    trainJob={job}
-                    onInterrupt={() => onInterrupt(job)}
-                    onViewLogs={() => onViewLogs(job)}
-                />
-            ))}
-        </View>
-    );
-};
-
-const useProjectTrainingJobs = (project_id: string): SchemaTrainJob[] => {
-    const { data: allJobs = [] } = $api.useQuery('get', '/api/jobs');
-
-    return allJobs
-        .filter((job) => job.project_id === project_id)
-        .filter((job): job is SchemaTrainJob => job.type === 'training');
-};
-
-const NoModelsPlaceholder = () => {
-    return (
-        <Flex margin={'size-200'} direction={'column'} height='100%'>
-            <IllustratedMessage>
-                <EmptyIllustration />
-                <Content> Currently there are no trained models available. </Content>
-                <Text>If you&apos;ve recorded a dataset it&apos;s time to begin training your model. </Text>
-                <Heading>No trained models</Heading>
-                <View margin={'size-100'}>
-                    <DialogTrigger>
-                        <Button variant='accent'>Train model</Button>
-                        {(close) => <TrainModelDialog close={close} />}
-                    </DialogTrigger>
-                </View>
-            </IllustratedMessage>
-        </Flex>
-    );
-};
-
-export const Index = () => {
-    const { project_id } = useProjectId();
-    const { data: models } = $api.useSuspenseQuery('get', '/api/projects/{project_id}/models', {
-        params: { path: { project_id } },
-    });
-
-    const jobs = useProjectTrainingJobs(project_id);
-    const [retrainModel, setRetrainModel] = useState<SchemaModel | null>(null);
-    const [logsSourceId, setLogsSourceId] = useState<string | undefined>();
-
-    const handleViewLogs = (model: SchemaModel) => {
-        if (!model.train_job_id) {
-            return;
-        }
-
-        setLogsSourceId(model.train_job_id);
-    };
-
-    const {} = useWebSocket(fetchClient.PATH('/api/jobs/ws'), {
-        shouldReconnect: () => true,
-        onMessage: (event: WebSocketEventMap['message']) => onMessage(event),
-    });
-    const client = useQueryClient();
-
-    const updateJob = (job: SchemaJob) => {
-        client.setQueryData<SchemaJob[]>(['get', '/api/jobs'], (old = []) => {
-            return old.map((m) => (m.id === job.id ? job : m));
-        });
-    };
-
-    const addJob = (job: SchemaJob) => {
-        client.setQueryData<SchemaJob[]>(['get', '/api/jobs'], (old = []) => {
-            return [...old, job];
-        });
-    };
-
-    const onMessage = ({ data }: WebSocketEventMap['message']) => {
-        const message_data = JSON.parse(data);
-        if (message_data.event === 'JOB_UPDATE') {
-            const message = message_data as { event: string; data: SchemaJob };
-            if (message.data.project_id !== project_id) {
-                return;
-            }
-
-            updateJob(message.data as SchemaTrainJob);
-
-            if (message.data.message && message.data.status === 'running') {
-                notify('info', message.data.message);
-            }
-
-            if (message.data.status === 'completed') {
-                client.invalidateQueries({ queryKey: ['get', '/api/projects/{project_id}/models'] });
-            }
-        }
-    };
-
-    const hasModels = models.length > 0;
-    const hasJobs = jobs.length > 0;
-    const showIllustratedMessage = !hasModels && !hasJobs;
-
-    return (
-        <View height='100%' padding={'size-300'} UNSAFE_style={{ overflowY: 'auto' }}>
-            <Flex direction={'column'} height={'100%'}>
-                {showIllustratedMessage ? (
-                    <NoModelsPlaceholder />
-                ) : (
-                    <Flex direction={'column'} flex={1} gap={'size-300'}>
-                        <Flex justifyContent={'end'}>
-                            <DialogTrigger>
-                                <Button variant='accent'>Train model</Button>
-                                {(close) => (
-                                    <TrainModelDialog
-                                        close={(job) => {
-                                            if (job) addJob(job);
-                                            close();
-                                        }}
-                                    />
-                                )}
-                            </DialogTrigger>
-                        </Flex>
-                        <JobList
-                            jobs={jobs}
-                            onViewLogs={(job) => {
-                                setLogsSourceId(job.id);
-                            }}
-                        />
-                        {hasModels && (
-                            <ModelList
-                                models={models}
-                                jobs={jobs}
-                                onRetrain={setRetrainModel}
-                                onViewLogs={handleViewLogs}
-                            />
-                        )}
-                    </Flex>
-                )}
-            </Flex>
-            <DialogContainer onDismiss={() => setRetrainModel(null)}>
-                {retrainModel && (
-                    <TrainModelDialog
-                        baseModel={retrainModel}
-                        close={(job) => {
-                            if (job) addJob(job);
-                            setRetrainModel(null);
-                        }}
-                    />
-                )}
-            </DialogContainer>
-            <DialogContainer type='fullscreen' onDismiss={() => setLogsSourceId(undefined)}>
-                {logsSourceId != null && (
-                    <LogsDialog close={() => setLogsSourceId(undefined)} initialSourceId={`job-${logsSourceId}`} />
-                )}
-            </DialogContainer>
-        </View>
-    );
-};
+export { ModelsPage as Index } from '../../features/models/models-page';
