--- application/ui/src/routes/models/utils.ts
@@ -1,14 +1,0 @@
-export const formatDuration = (ms: number): string => {
-    const totalSeconds = Math.floor(ms / 1000);
-    const hours = Math.floor(totalSeconds / 3600);
-    const minutes = Math.floor((totalSeconds % 3600) / 60);
-    const seconds = totalSeconds % 60;
-
-    if (hours > 0) return `${hours}h ${minutes}m ${seconds}s`;
-    if (minutes > 0) return `${minutes}m ${seconds}s`;
-    return `${seconds}s`;
-};
-
-export const durationBetween = (start: string, end: string): string => {
-    return formatDuration(new Date(end).getTime() - new Date(start).getTime());
-};
