+++ application/ui/src/utils.ts
@@ -13,3 +13,14 @@ const pluralRules = new Intl.PluralRules('en');
 
 export const pluralize = (count: number, singular: string, plural: string): string =>
     pluralRules.select(count) === 'one' ? singular : plural;
+
+export const formatDuration = (ms: number): string => {
+    const totalSeconds = Math.floor(ms / 1000);
+    const hours = Math.floor(totalSeconds / 3600);
+    const minutes = Math.floor((totalSeconds % 3600) / 60);
+    const seconds = totalSeconds % 60;
+
+    if (hours > 0) return `${hours}h ${minutes}m ${seconds}s`;
+    if (minutes > 0) return `${minutes}m ${seconds}s`;
+    return `${seconds}s`;
+};
